Update README.md
For whatever reason, GitHub's Markdown parser is not particularly bright, so it messes up the rendering of some LaTeX input because it sanitizes Markdown and probably first renders it as html, and only applies the MathJax engine at the end of the process. With trial and error, adding a whitespace before closing the math expression with a dollar sign seems to usually do the trick. See https://nschloe.github.io/2022/05/20/math-on-github.html